Task 1. Fixed problem. Improve UI solution.
